Move utils and samples in a dedicated directory
