--- Samples/Samples.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27004.2008
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PAA2PNG", "PAA2PNG\PAA2PNG.csproj", "{AA61F860-E10E-4B5C-9AC0-EA92E58F2864}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BIS.PAA", "..\BIS.PAA\BIS.PAA.csproj", "{389D8D45-86C9-4804-B73D-21551203A36B}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BIS.Core", "..\BIS.Core\BIS.Core.csproj", "{2120BBD0-EA89-41E8-B63B-ED04F7720BF9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AA61F860-E10E-4B5C-9AC0-EA92E58F2864}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AA61F860-E10E-4B5C-9AC0-EA92E58F2864}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AA61F860-E10E-4B5C-9AC0-EA92E58F2864}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AA61F860-E10E-4B5C-9AC0-EA92E58F2864}.Release|Any CPU.Build.0 = Release|Any CPU
-		{389D8D45-86C9-4804-B73D-21551203A36B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{389D8D45-86C9-4804-B73D-21551203A36B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{389D8D45-86C9-4804-B73D-21551203A36B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{389D8D45-86C9-4804-B73D-21551203A36B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2120BBD0-EA89-41E8-B63B-ED04F7720BF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2120BBD0-EA89-41E8-B63B-ED04F7720BF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2120BBD0-EA89-41E8-B63B-ED04F7720BF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2120BBD0-EA89-41E8-B63B-ED04F7720BF9}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A1FE6AFC-9DE0-4EFC-A294-BEC6436D1513}
-	EndGlobalSection
-EndGlobal
