+++ bis-file-formats.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.31321.278
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bis-file-formats", "bis-file-formats\bis-file-formats.csproj", "{4CB31AF5-DD6D-4547-BAEF-092AB3D06CCD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WrpUtil", "WrpUtil\WrpUtil.csproj", "{184D5A1E-8A5A-4E15-AF39-19EBB72C80BC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WrpUtil", "Utils\WrpUtil\WrpUtil.csproj", "{184D5A1E-8A5A-4E15-AF39-19EBB72C80BC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PAA2PNG", "Utils\PAA2PNG\PAA2PNG.csproj", "{AA61F860-E10E-4B5C-9AC0-EA92E58F2864}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{184D5A1E-8A5A-4E15-AF39-19EBB72C80BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{184D5A1E-8A5A-4E15-AF39-19EBB72C80BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{184D5A1E-8A5A-4E15-AF39-19EBB72C80BC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AA61F860-E10E-4B5C-9AC0-EA92E58F2864}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AA61F860-E10E-4B5C-9AC0-EA92E58F2864}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AA61F860-E10E-4B5C-9AC0-EA92E58F2864}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AA61F860-E10E-4B5C-9AC0-EA92E58F2864}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
